requirements.txt
flask==2.0.1
beautifulsoup4
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
transformers==4.36.2
datasets==2.16.1
librosa==0.10.1
evaluate==0.4.1
jiwer==3.0.3
numpy==1.24.3
pandas==2.0.2
tqdm==4.65.0
accelerate==0.25.0
wandb==0.15.4
soundfile==0.12.1
scipy==1.11.4
